+++ app/src/main/java/app/hillforts/activities/HillfortListActivity.kt
@@ -10,16 +10,19 @@ import org.jetbrains.anko.startActivityForResult
 import app.hillforts.R
 import app.hillforts.main.MainApp
 import app.hillforts.models.HillfortModel
+import app.hillforts.models.UserModel
 import org.jetbrains.anko.intentFor
 
 class HillfortListActivity : AppCompatActivity(), HillfortListener {
 
     lateinit var app: MainApp
+    lateinit var user: UserModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_hillfort_list)
         app = application as MainApp
+        user = app.appUser
 
         toolbar.title = title
         setSupportActionBar(toolbar)
@@ -30,10 +33,10 @@ class HillfortListActivity : AppCompatActivity(), HillfortListener {
     }
 
     private fun loadHillforts() {
-        showHillforts(app.hillforts.findAll())
+        showHillforts(this.app.unified.findAllHillfortsForUser(app.appUser))
     }
 
-    fun showHillforts (hillforts: List<HillfortModel>) {
+    private fun showHillforts (hillforts: List<HillfortModel>) {
         recyclerView.adapter = HillfortAdapter(hillforts, this)
         recyclerView.adapter?.notifyDataSetChanged()
     }