+++ app/src/main/java/app/hillforts/activities/LoginActivity.kt
@@ -8,15 +8,25 @@ import kotlinx.android.synthetic.main.activity_login.*
 import android.text.method.PasswordTransformationMethod
 import android.widget.Button
 import android.widget.EditText
+import app.hillforts.main.MainApp
+import app.hillforts.models.UserModel
+import app.hillforts.models.HillfortModel
+import org.jetbrains.anko.toast
 
 class LoginActivity : AppCompatActivity() {
+
+    lateinit var app: MainApp
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
+        app = application as MainApp
 
+        var user = UserModel()
+        var password: EditText = findViewById(R.id.input_password)
+        var email: EditText = findViewById(R.id.input_email)
         var showHideBtn:Button = findViewById(R.id.showHideBtn)
-        var password:EditText = findViewById(R.id.input_password)
+        val users = app.unified.findAllUsers()
 
         showHideBtn.setOnClickListener {
             if (showHideBtn.getText().toString().equals("Show Password")) {
@@ -29,9 +39,17 @@ class LoginActivity : AppCompatActivity() {
         }
 
         btnLogin.setOnClickListener() {
-            val intent = Intent(applicationContext, HillfortListActivity::class.java)
-            startActivity(intent)
-            finish()
+            user.email = email.text.toString()
+            user.password = password.text.toString()
+            for (i in users) {
+                if (i.email == user.email && i.password == user.password) {
+                    app.appUser = i
+                    val intent = Intent(applicationContext, HillfortListActivity::class.java)
+                    startActivity(intent)
+                    finish()
+                }
+            }
+
         }
 
         linkSignup.setOnClickListener() {